requirements.txt
tavily-python>=0.5.0
anthropic>=0.40.0
requests>=2.32.3
python-dotenv>=1.0.1
